backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
fpdf2==2.8.2
